Feature dagster (#507)
* implement load_fraud_data

* add transform-fraud-data asset

* initial commit

* [fix]

* fix: adjust text overlay position well
