+++ src/students/HayBeeCoder/gradioapp/image_app.py
@@ -1,7 +1,4 @@
-try:
-    from .ui import create_interface
-except ImportError:
-    from ui import create_interface
+from .ui import create_interface
 
 # Create the main application interface
 app = create_interface()
