+++ src/students/HayBeeCoder/gradioapp/text.py
@@ -6,7 +6,8 @@ except ImportError:
     from image_utils import load_font, parse_color
 
 
-def add_text_overlay(img, text, color=(0, 0, 0), font_size=50, position="center", font_type="dejavusans"):
+def add_text_overlay(img, text, color=(0, 0, 0), font_size=50, position="center",
+                      font_type="dejavusans"):
     """Add text overlay to image."""
     if not text:
         return img
@@ -26,13 +27,13 @@ def add_text_overlay(img, text, color=(0, 0, 0), font_size=50, position="center"
     if position == "center":
         pos = ((img.width - w) // 2, (img.height - h) // 2)
     elif position == "top_left":
-        pos = (0, 0)
+        pos = (20, 0)
     elif position == "top_right":
-        pos = (img.width - w, 0)
+        pos = (img.width - w - 20, 0)
     elif position == "bottom_left":
-        pos = (0, img.height - h)
+        pos = (0 + 10, img.height - h - 15)
     elif position == "bottom_right":
-        pos = (img.width - w, img.height - h)
+        pos = (img.width - w - 15, img.height - h - 15)
     else:
         pos = ((img.width - w) // 2, (img.height - h) // 2)
